[eclipse/xtext#1679] converted Xtend code to Java
